DiscordAIBot/requirements.txt
discord.py
python-dotenv
requests
json
flask